chatbot/django/requirements.txt
django
django-browser-reload
django-extras
django-extensions
django-debug-toolbar
django-tailwind[reload]
channels
daphne
django-htmx
langchain-ollama
langchain-chroma
langchain
langchain-community
pytest-django
psycopg2-binary
websocket-client
whitenoise
randomname
djangorestframework
django-custom-user
django-sql-explorer
dj-rest-auth
django-cors-headers
djangorestframework-simplejwt
django-environ
